refactor: restructure before implementing more search functions
