+++ src/core/controllers/searching/CMakeLists.txt
@@ -1,0 +1,4 @@
+target_sources(${DND_CORE}
+    PRIVATE
+    
+)
