+++ src/core/controllers/searching/trie.hpp
